// openpos-client-libs/projects/openpos-client-core-lib/src/lib/shared/screen-parts/prompt-form-part-card/prompt-form-part-card.interface.ts
import { FieldInputType } from '../../../core/interfaces/field-input-type.enum';
import { IActionItem } from '../../../core/actions/action-item.interface';
import { Validator } from '@angular/forms';
import { PromptPosition } from './prompt-position-card.enum';
import { DisplayProperty } from '../../components/display-property/display-property.interface';
import { ScanInterface } from '../scan-part/scan-part.interface';

export interface PromptFormPartCardInterface {
    type: string;
    promptIcon: string;
    placeholderText: string;
    hintText: string;
    instructions: string;
    responseText: string;
    editable: boolean;
    responseType: FieldInputType;
    actionButton: IActionItem;
    secondaryActionButton: IActionItem;
    otherActions: IActionItem[];
    minLength: number;
    maxLength: number;
    max: number;
    min: number;
    validationPatterns: Array<string>;
    scanEnabled: boolean;
    keyboardPreference: string;
    isRequiredInputField: boolean;
    isGiftCardScanEnabled: boolean;
    validators: Validator[];
    validationMessages: Map<string, string>;
    promptPosition: PromptPosition;
    info: DisplayProperty[];
    scan?: ScanInterface;
    autoFocus: boolean;
    hintDateText: string;
    isExpiryDateEnabled: boolean;
    todayDate: Date;
    placeholderDateText: string;
    responseDateText: string;
    responseDateType: FieldInputType;
}